feat!: split the CLI entry from the SDK library entry
src/index.ts was both the bin and the library root and re-exported the
whole CLI, so SDK consumers loaded Commander and ~30 command classes at
import time. The bin is now a dedicated src/cli.ts (dist/cli.js) that
runs unconditionally, so the shouldRunMain guard machinery is gone; the
root entry exports only the SDK surface, and the command classes remain
available to embedders via the new portfolio-manager/cli export path.

BREAKING CHANGE: CLI command classes are no longer exported from the
package root; import them from portfolio-manager/cli instead. The
shouldRunMain(importMetaUrl, argv1) helper is removed entirely — it
existed only so the combined entry could decide whether to run the CLI,
which the dedicated bin makes moot. The bin path moved from
dist/index.js to dist/cli.js (the installed portfolio-manager command
is unaffected).
