+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
       - run: npm run format:check
       - run: npm run typecheck
       - run: npm run build
-      - run: node ./dist/index.js --help # Run the CLI to ensure it works, we had some test releases where this was broken during the esm migration
+      - run: node ./dist/cli.js --help # Run the CLI to ensure it works, we had some test releases where this was broken during the esm migration
       - run: npm test
 
   release:
@@ -72,7 +72,7 @@ jobs:
       - run: npm ci
       - run: npm run typecheck
       - run: npm run build
-      - run: node ./dist/index.js --help # Run the CLI to ensure it works, we had some test releases where this was broken during the esm migration
+      - run: node ./dist/cli.js --help # Run the CLI to ensure it works, we had some test releases where this was broken during the esm migration
       - run: npm run test:coverage # Run tests with coverage to ensure we fail if there are uncovered lines (e.g. due to a new file missing tests)
       - name: Semantic Release with Trusted Publishing via OIDC
         env:
