+++ package.json
@@ -9,7 +9,7 @@
     "dist/"
   ],
   "bin": {
-    "portfolio-manager": "dist/index.js"
+    "portfolio-manager": "dist/cli.js"
   },
   "type": "module",
   "main": "dist/index.js",
@@ -18,6 +18,10 @@
     ".": {
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./cli": {
+      "types": "./dist/cli/index.d.ts",
+      "import": "./dist/cli/index.js"
     },
     "./package.json": "./package.json"
   },
@@ -34,8 +38,8 @@
     "typecheck:e2e": "tsc --noEmit -p tsconfig.e2e.json",
     "wipe:test-environment": "tsx ./scripts/wipeTestEnvironment.ts",
     "semantic-release": "semantic-release",
-    "src-bin": "tsx ./src/index.ts",
-    "build-bin": "node ./dist/index.js"
+    "src-bin": "tsx ./src/cli.ts",
+    "build-bin": "node ./dist/cli.js"
   },
   "author": {
     "name": "Darrel O'Pry"
